Fixed incorrect column type

+++ db/migration/000001_init_schema.up.sql
@@ -13,7 +13,7 @@ CREATE TABLE "campos" (
   "titulo" varchar NOT NULL,
   "tipo" varchar NOT NULL,
   "hectareas" bigint NOT NULL,
-  "precio_por_hectarea" float NOT NULL,
+  "precio_por_hectarea" bigint NOT NULL,
   "ciudad" varchar NOT NULL,
   "provincia" varchar NOT NULL,
   "fecha_creacion" timestamptz NOT NULL DEFAULT (now())
